+++ src/utils/log.ts
@@ -1,0 +1,3 @@
+import chalk from 'chalk';
+
+export const error = (...args: unknown[]) => console.error('❌', chalk.bold.red('Error'), ...args);
